Generate socket-actions
