+++ webofneeds/won-owner-webapp/src/main/webapp/style/_atom-footer.scss
@@ -11,6 +11,7 @@ won-atom-footer {
     text-align: left;
   }
 
+  won-socket-actions,
   .atom-footer__buttons {
     display: grid;
     grid-gap: 0.5rem;
